+++ example/rails-60/app/controllers/que_controller.rb
@@ -1,0 +1,14 @@
+require './app/jobs/que_crash'
+require './app/jobs/que_callback'
+
+class QueController < ActionController::Base
+  layout "application"
+
+  def crash
+    QueCrash.enqueue
+  end
+
+  def callbacks
+    QueCallback.enqueue
+  end
+end